Chatbubbles
Added chat bubbles to chat screen for sender + reciever. Automatically add new chat to stack when message is sent. Messages that are sent are not immediately added to local stack. Buffer for local messages not sent yet needs to be implemented. Added connectedClients map to route message to correct socket based on ID (email). Chats are automatically sorted with newest chats at the top.